[chore] move redis folder into servers
